requirements.txt
fastapi
sqlmodel
supabase
python-dotenv
psycopg2-binary
python-jose[cryptography]

# dev
ipython
